Renamed package *.comment.* to *.modifiertags.*
