+++ test-app-filters/test/be/objectify/deadbolt/java/test/controllers/modifiertags/CommentCompositeConstraintsTest.java
@@ -13,14 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package be.objectify.deadbolt.java.test.controllers.comment;
+package be.objectify.deadbolt.java.test.controllers.modifiertags;
 
-import be.objectify.deadbolt.java.test.controllers.RoleBasedPermissionsTest;
+import be.objectify.deadbolt.java.test.controllers.CompositeConstraintsTest;
 
 /**
  * @author Steve Chaloner (steve@objectify.be)
  */
-public class CommentRoleBasedPermissionsTest extends RoleBasedPermissionsTest
+public class CommentCompositeConstraintsTest extends CompositeConstraintsTest
 {
     @Override
     public String pathComponent()
