+++ test-app-filters/test/be/objectify/deadbolt/java/test/controllers/modifiertags/CommentPatternEqualityTest.java
@@ -13,14 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package be.objectify.deadbolt.java.test.controllers.comment;
+package be.objectify.deadbolt.java.test.controllers.modifiertags;
 
-import be.objectify.deadbolt.java.test.controllers.CompositeConstraintsTest;
+import be.objectify.deadbolt.java.test.controllers.PatternEqualityTest;
 
 /**
  * @author Steve Chaloner (steve@objectify.be)
  */
-public class CommentCompositeConstraintsTest extends CompositeConstraintsTest
+public class CommentPatternEqualityTest extends PatternEqualityTest
 {
     @Override
     public String pathComponent()
