+++ test-app-filters/test/be/objectify/deadbolt/java/test/controllers/modifiertags/CommentPatternRegexTest.java
@@ -13,14 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package be.objectify.deadbolt.java.test.controllers.comment;
+package be.objectify.deadbolt.java.test.controllers.modifiertags;
 
-import be.objectify.deadbolt.java.test.controllers.SubjectNotPresentTest;
+import be.objectify.deadbolt.java.test.controllers.PatternRegexTest;
 
 /**
  * @author Steve Chaloner (steve@objectify.be)
  */
-public class CommentSubjectNotPresentTest extends SubjectNotPresentTest
+public class CommentPatternRegexTest extends PatternRegexTest
 {
     @Override
     public String pathComponent()
