+++ test-app-filters/test/be/objectify/deadbolt/java/test/controllers/modifiertags/CommentRestrictTest.java
@@ -13,14 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package be.objectify.deadbolt.java.test.controllers.comment;
+package be.objectify.deadbolt.java.test.controllers.modifiertags;
 
-import be.objectify.deadbolt.java.test.controllers.PatternRegexTest;
+import be.objectify.deadbolt.java.test.controllers.RestrictTest;
 
 /**
  * @author Steve Chaloner (steve@objectify.be)
  */
-public class CommentPatternRegexTest extends PatternRegexTest
+public class CommentRestrictTest extends RestrictTest
 {
     @Override
     public String pathComponent()
