+++ test-app-filters/test/be/objectify/deadbolt/java/test/controllers/modifiertags/CommentSubjectNotPresentTest.java
@@ -13,14 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package be.objectify.deadbolt.java.test.controllers.comment;
+package be.objectify.deadbolt.java.test.controllers.modifiertags;
 
-import be.objectify.deadbolt.java.test.controllers.SubjectPresentTest;
+import be.objectify.deadbolt.java.test.controllers.SubjectNotPresentTest;
 
 /**
  * @author Steve Chaloner (steve@objectify.be)
  */
-public class CommentSubjectPresentTest extends SubjectPresentTest
+public class CommentSubjectNotPresentTest extends SubjectNotPresentTest
 {
     @Override
     public String pathComponent()
